requirements.txt
pyyaml>=6.0
requests>=2.31.0
streamlit>=1.30.0

